.cosmetics
git-svn-id: svn://ultimatepp.org/upp/trunk@13705 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/HexView/HexView.cpp
@@ -291,7 +291,7 @@ void HexView::SetCursor(uint64 _cursor)
 	if(cursor < sc) {
 		sc = cursor;
 	}
-	if(sc > q)
+	if(sc > (uint64)q)
 		sc = (sc - q) / columns * columns + q;
 	if(sc >= total)
 		sc = total - 1;
